requirements.txt
pyFirmata
pyserial
pyglet
coloredlogs